Updated to only show campusses when provider has none and also replaced slider with add + remove buttons (ford3-162) (#170)
 Updated to only show campusses when provider has none and also replaced slider with add + remove buttons (ford3-162) (#170)
