+++ django_project/ford3/static/js/provider_form.js
@@ -1,34 +1,50 @@
-var slider = document.getElementById('campus-count');
-var output = document.getElementById('number-of-campuses');
-slider.value = 1;
-output.value = slider.value; // Display the default slider value
-// Update the current slider value (each time you drag the slider handle)
-slider.oninput = function () {
-    output.value = this.value;
+$(document).ready(function() {
+    innitiateRemoveCampusButtons();
+})
 
-    createCampusNameInputs(this.value);
-}
+function innitiateRemoveCampusButtons() {
+    $('.remove-campus-button').click(function() {
+        this_button_parrent_parent = $(this).parent().parent();
+        this_button_parrent_parent.remove();
+    })
+ }
+
+$(document).on("submit", "form", function(e) {
+    if (confirm('Submitting data. Continue?')) {
+        campus_name_elements = $('.campus_name').elements
+       return true;
+    }
+    else {
+       return false;
+    }
+})
+
+$('#add-campus-name').click(
+    function () {
+        addCampusNameInput();
+        innitiateRemoveCampusButtons();
+    }
+ )
 
 
-function createCampusNameInputs(campus_count) {
-    var i;
-    var result_html = '';
-    var campus_container = document.getElementById('campus-names-input-wrapper');
-    var campus_count_number;
-    campus_container.innerHTML = '';
-    for (i = 0; i < campus_count; i++) {
-        campus_count_number = i + 1;
-        result_html +=
+
+function addCampusNameInput() {
+    var campus_container = $('#campus-names-input-wrapper');
+    var result_html = ('<div>' +
             '<div class="row mt1 campus-name-fade-in">' +
                 '<div class="col-4 centerv npl">' +
-                    'Name of Campus ' + campus_count_number.toString() +
+                    'Campus Name ' +
                 '</div>' +
-                '<div class="col-8">' +
-                    '<input name="campus_name_' + campus_count_number.toString() + '" ' +
-                    'id="campus_name_' + campus_count_number.toString() + '" ' +
+                '<div class="col-7">' +
+                    '<input name="campus_name" required class="campus_name" ' +
                     'type="text" placeholder="•••••••••••••••••"/>' +
                 '</div>' +
-            '</div>'
-    }
-    campus_container.innerHTML = result_html;
+                '<div class="col-1">' +
+                    '<div class="remove-campus-button">' +
+                    '<div class="remove-campus-button-inner ">X</div></div>' +
+                '</div>' +
+            '</div></div>')
+    new_input = $.parseHTML(result_html);
+    campus_container.append(new_input);
+
 }
